pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "xrpl-hermes"
version = "1.3.8"
description = "Open-source XRPL ecosystem toolkit — 59 knowledge files + 64 working tools covering L1, EVM Sidechain, Xahau Hooks, Flare FTSO, Axelar Bridge, Arweave"
authors = [{name = "CarpXRPL"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["xrpl-py>=2.5.0", "httpx>=0.25.0"]

[tool.setuptools.packages.find]
include = ["scripts*"]
